refactor: rename SnakesAndLadders to LaddersGame.
